webdetective/backend/requirements.txt
flask
flask-cors
requests
beautifulsoup4
pandas
nltk
torch
faiss-cpu
numpy
tqdm
transformers
sentence-transformers

